requirements.txt
# LangChain ecosystem (fixed versions)
langchain-core==0.1.16
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13
langsmith==0.0.87
langgraph==0.0.24

# FastAPI and frontend
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
streamlit==1.29.0

# Auth & Google APIs
google-auth>=2.6.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Embeddings & vector DB
chromadb>=0.3.0
sentence-transformers==2.2.2
pymongo>=4.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
tiktoken>=0.3.0

# HTTP
httpx>=0.24.0
requests>=2.26.0

# Config and parsing
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-core>=2.14.1,<3.0.0
pydantic-settings>=2.0.0

# Logging
structlog==23.2.0
python-json-logger>=2.0.0

# OpenAI
openai>=1.0.0

# Utilities
PyJWT>=2.3.0
cryptography>=3.4.0
